# api/schema/openapi.yaml
openapi: 3.1.0
info:
  title: CMCEN / RCMCE API
  version: 1.0.0
  description: >
    OpenAPI-style inventory for the Express routes mounted by server/server.js.
    Schemas are intentionally broad where handlers return Mongoose documents.
    The existing developer actor may save blank descriptive profile/editorial
    fields. Ordinary actors retain required checks but may retain authorized
    blanks on later edits. Credentials, identifiers, references, valid nonblank
    values, dates and publication requirements remain enforced; see API ROUTES.md.
    Every /api endpoint is rate limited by source IP and may return 429 with
    Retry-After and RateLimit-* headers. Sensitive endpoints have stricter limits.
    Every response includes X-Request-ID for support correlation. Server failures
    are logged and audited without retaining request bodies or query strings.
servers:
  - url: http://localhost:3000
    description: Local development (Eric)
  - url: http://localhost:3001
    description: Local development (Bray)
  - url: https://cmcen-staging.corebot.ca
    description: Staging
  - url: https://beta.cmcen-rcmce.ca
    description: Beta
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/comments/{commentId}:
    patch:
      tags: [Admin, Retirements]
      summary: Edit a comment.
      description: Requires canReviewAndPublish. Corrects the comment body without changing its lifecycle state. Changes are audited.
      x-permission: canReviewAndPublish
      parameters:
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        $ref: "#/components/requestBodies/GenericJson"
      responses:
        "200":
          description: Comment updated.
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Admin, Retirements]
      summary: Delete any comment.
      x-permission: canDeleteContent
      parameters:
        - $ref: "#/components/parameters/CommentId"
      responses:
        "200":
          description: Delete result.
...
  /api/comments/{commentId}/review:
    patch:
      tags: [Retirements]
      summary: Publish or reject comment.
      description: Reviewers may publish draft or pending comments. Rejection applies only to pending submissions. Draft edits use the admin comments endpoint and preserve status; personal submission edits return 409 for drafts.
      x-permission: canReviewAndPublish
      parameters:
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        $ref: "#/components/requestBodies/ImmediateReviewDecision"
      responses:
        "200":
          description: Review result.
  /api/comments/{commentId}/edit:
    get:
      tags: [Retirements]
      summary: Get comment edit payload.
      parameters:
        - $ref: "#/components/parameters/CommentId"
      responses:
        "200":
          description: Comment edit payload.
  /api/comments/{commentId}:
    patch:
      tags: [Retirements]
      summary: Update comment.
      description: Optional submitForReview=true restricts resubmission to the owner of a pending/rejected comment and keeps it pending regardless of staff publishing permissions. Existing behavior is retained when omitted.
      parameters:
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 2, maxLength: 10000 }
                submitForReview: { type: boolean, default: false }
      responses:
        "200":
          description: Updated comment.
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/comments/on/{parentType}/{parentId}:
    get:
      tags: [Comments]
      summary: List published comments for message.
      security: []
      parameters:
        - name: parentType
          in: path
          required: true
          description: A post type registered in config/comment-targets.js.
          schema: { type: string }
        - name: parentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Comments.
    post:
      tags: [Comments]
      summary: Create comment.
      parameters:
        - name: parentType
          in: path
          required: true
          description: A post type registered in config/comment-targets.js.
          schema: { type: string }
        - name: parentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        $ref: "#/components/requestBodies/GenericJson"
      responses:
        "201":
          description: Created comment.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CommentId:
      name: commentId
      in: path
      required: true
      schema: { type: string }
...
  requestBodies:
    GenericJson:
      required: false
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...
    ImmediateReviewDecision:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [action]
            properties:
              action:
                type: string
                enum: [publish, reject]
              publicationDateChoice:
                type: string
                enum: [original, now, custom]
                description: Required for imported/archive comments. Original uses the stored original public date; originally unapproved comments cannot use original. Invalid or unavailable choices return 400.
              customPublishedAt:
                type: string
                format: date-time
                description: Required with custom publicationDateChoice; cannot be in the future.
              rejectionReason:
                type: string
                maxLength: 2000
                description: Required when `action` is `reject`.
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...